renamed raw_stream.py -> raw.py, initial_stream.py -> initial.py
