+++ tests/test_stream_app.py
@@ -6,7 +6,7 @@ from corva.application import Corva
 from corva.configuration import SETTINGS
 from corva.models.stream.context import BaseStreamContext
 from corva.models.stream.log_type import LogType
-from corva.models.stream.raw_stream import (
+from corva.models.stream.raw import (
     RawAppMetadata,
     RawDepthRecord,
     RawMetadata,
